requirements.txt
motor
pillow
faker
reportlab